cleaned up some code and changed some namings
