+++ code/src/components/AddTodoButton.js
@@ -2,9 +2,7 @@ import React from 'react'
 
 import assets from '../assets/icons/icons8-plus-48.png'
 
-const AddButton = ({ isAddInputFilled }) => {
-
-
+const AddTodoButton = ({ isAddInputFilled }) => {
   if (isAddInputFilled()) {
     return (
       <button className="add-button" type="submit">
@@ -13,11 +11,10 @@ const AddButton = ({ isAddInputFilled }) => {
     );
   } else {
     return (
-
       <button className="add-button" type="submit" disabled>
         <img src={assets} alt="A plus icon" />
       </button>
     )
   }
 }
-export default AddButton
+export default AddTodoButton
